+++ packages/crdt_socket_sync/lib/src/plugins/common/plugin.dart
@@ -1,9 +1,8 @@
 import 'package:crdt_socket_sync/src/common/common.dart';
 
-/// Base class for all server plugins.
+/// Base class for all server and client plugins.
 ///
-/// A plugin can be used to extend the server functionality.
-/// It can handle custom messages, and listen to events from the server.
+/// A plugin can be used to extend the server or client functionality.
 abstract class SyncPlugin {
   /// The name of the plugin.
   String get name;
@@ -11,11 +10,10 @@ abstract class SyncPlugin {
   /// The message codec for the plugin.
   ///
   /// This is used to encode and decode messages for the plugin.
-  /// The default implementation is [CompressedCodec] with [NoCompression].
   MessageCodec<Message> get messageCodec;
 
   /// Called when the plugin is disposed.
   ///
-  /// Usually [dispose] is called when the server is disposed.
+  /// Usually [dispose] is called when the server or client is disposed.
   void dispose();
 }