+++ packages/crdt_socket_sync/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [Unreleased]
+
+### Added
+- Feature: add plugin system
+- Feature: add awareness plugin
+
 ## [0.1.0+1] - 2025-06-14
 ### Fixed
 - Chore: update readme links